feat: fix EOS detection, add jax.nn.dot_product_attention, and JIT sampling
- Support multiple EOS token IDs from generation_config.json (GLM-4 has 3)
- Replace manual attention with jax.nn.dot_product_attention (native GQA)
- Add fp32 softmax in MLA attention for numerical stability
- Add sampling with temperature, top_p, top_k, and repetition penalty
- JIT-compile sampler with top_k pre-filtering (155K→200 vocab sort)
- Add --sample flag and sampling parameters to main.py CLI
- Document full debugging process in docs/glm4_repetition_debugging.md

+++ layers/mla_attention.py
@@ -182,7 +182,7 @@ class MLAAttention(nnx.Module):
         if attention_mask is not None:
             attn_weights = attn_weights + attention_mask
 
-        attn_weights = jax.nn.softmax(attn_weights, axis=-1).astype(v_full.dtype)
+        attn_weights = jax.nn.softmax(attn_weights.astype(jnp.float32), axis=-1).astype(v_full.dtype)
         attn_output = jnp.matmul(attn_weights, v_full)  # [B, H, S, v_head_dim]
 
         # Reshape back: [B, H, S, v] → [B, S, H*v]
